pass SNAPSHOT into the deploy-iqe script
