+++ files/bin/models.py
@@ -1,0 +1,53 @@
+from collections.abc import MutableMapping
+from typing import Any
+
+from pydantic import AnyUrl
+from pydantic import BaseModel
+from pydantic import ConfigDict
+from pydantic import Field
+from pydantic import model_validator
+
+
+class Git(BaseModel):
+    model_config = ConfigDict(frozen=True)
+
+    url: AnyUrl
+    revision: str
+
+
+class Source(BaseModel):
+    model_config = ConfigDict(frozen=True)
+
+    git: Git
+
+
+class ContainerImage(BaseModel):
+    model_config = ConfigDict(frozen=True)
+
+    image: str
+    sha: str
+
+
+class Component(BaseModel):
+    model_config = ConfigDict(frozen=True)
+
+    name: str
+    container_image: ContainerImage = Field(alias="containerImage")
+    source: Source
+
+    @model_validator(mode="before")
+    @classmethod
+    def container_image_validator(cls, data: Any) -> Any:
+        if not isinstance(data, MutableMapping):
+            raise ValueError(f"{data} is not of mapping type")
+
+        image, sha = data["containerImage"].split("@sha256:")
+        data["containerImage"] = ContainerImage(image=image, sha=sha)
+        return data
+
+
+class Snapshot(BaseModel):
+    model_config = ConfigDict(frozen=True)
+
+    application: str
+    components: list[Component]
